requirements.txt
psycopg2-binary==2.9.9
elasticsearch==8.11.0
python-dotenv==1.0.0
PyYAML==6.0.1
tenacity==8.2.3
psycopg2-pool==1.1
ratelimit==2.2.1
fastapi==0.105.0
uvicorn==0.24.0
pydantic==2.5.2
